+++ lib/src/ui/impl/clipped_image_impl.dart
@@ -3,7 +3,13 @@
 
 part of dart_web_toolkit_ui;
 
-
+/**
+ * Uses a combination of a clear image and a background image to clip all except
+ * a desired portion of an underlying image.
+ *
+ * Do not use this class - it is used for implementation only, and its methods
+ * may change in the future.
+ */
 class ClippedImageImpl {
 
   String clearImage = DWT.getModuleBaseURL().concat("resource/images/clear.gif");
@@ -12,7 +18,6 @@ class ClippedImageImpl {
     dart_html.ImageElement img = new dart_html.ImageElement();
     img.src = clearImage;
     //
-    //String style = "url(\"${DWT.getModuleBaseURL().concat(url.asString())}\") no-repeat ${-left}px ${-top}px";
     String style = "url(\"${url.asString()}\") no-repeat ${-left}px ${-top}px";
     img.style.background = style;
     img.style.width = "${width}px";
@@ -25,12 +30,8 @@ class ClippedImageImpl {
   }
 
   SafeHtml getSafeHtml(SafeUri url, int left, int top, int width, int height) {
-
-    //String style = "url(\"${DWT.getModuleBaseURL().concat(url.asString())}\") no-repeat ${-left}px ${-top}px";
     String style = "url(\"${url.asString()}\") no-repeat ${-left}px ${-top}px";
-
     String res = "<img onload='this.data-${DomImpl.UNHANDLED_EVENT_ATTR}=\"load\";' src='${clearImage}' style='background:${style}; width: ${width}px; height: ${height}px' border='0'>";
-
     return new SafeHtmlString(res);
   }
 
@@ -39,7 +40,7 @@ class ClippedImageImpl {
   }
 
   void adjust(dart_html.ImageElement img, SafeUri url, int left, int top, int width, int height) {
-    String style = "url(\"${DWT.getModuleBaseURL().concat(url.asString())}\") no-repeat -${left}px -${top}px";
+    String style = "url(\"${url.asString()}\") no-repeat -${left}px -${top}px";
     img.style.background = style;
     img.style.width = width.toString().concat(Unit.PX.value);
     img.style.height = height.toString().concat(Unit.PX.value);